# openapi/glenkeos-complete-api-v1.yaml
openapi: 3.0.3
info:
  title: GlenKeos Platform API
  version: 1.0.0
  description: |
    Complete API specification for the GlenKeos multi-tenant commerce platform.
    Includes all 10 microservices, PayPal integration, and event-driven workflows.
  contact:
    name: GlenKeos Engineering
    email: engineering@glenkeos.com
servers:
  - url: https://api.glenkeos.com/v1
    description: Production
  - url: https://staging-api.glenkeos.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
security:
  - CognitoAuthorizer: []
components:
  securitySchemes:
    CognitoAuthorizer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
        timestamp:
          type: string
          format: date-time
    PaginationMeta:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer

    # Stores
    Store:
      type: object
      properties:
        id:
          type: string
          format: uuid
        brand_id:
          type: string
        name:
          type: string
        address:
          type: object
        status:
          type: string
          enum: [ACTIVE, INACTIVE, MAINTENANCE]
        capabilities:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time

    # Customers
...
paths:
...
  /stores:
    post:
      summary: Create a new store
      operationId: createStore
      tags: [Stores]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [brand_id, name, address]
              properties:
                brand_id:
                  type: string
                name:
                  type: string
                address:
                  type: object
                capabilities:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Store created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List stores
      operationId: listStores
      tags: [Stores]
      parameters:
        - name: brand_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of stores
          content:
            application/json:
              schema:
                type: object
                properties:
                  stores:
                    type: array
                    items:
                      $ref: '#/components/schemas/Store'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /stores/{storeId}:
    get:
      summary: Get store by ID
      operationId: getStore
      tags: [Stores]
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Store details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
        '404':
          description: Store not found

    put:
      summary: Update store
      operationId: updateStore
      tags: [Stores]
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: object
                status:
                  type: string
                capabilities:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Store updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'

  # ========================================
  # CUSTOMERS SERVICE
  # ========================================
...